+++ src/UI/operaWallet/helpers/getIsOperaWalletAvailable.ts
@@ -1,0 +1,3 @@
+export const getIsOperaWalletAvailable = () => {
+  return Boolean(window.isOpera && window.elrond);
+};