+++ src/hooks/login/index.ts
@@ -1,4 +1,5 @@
 export { useExtensionLogin } from './useExtensionLogin';
+export { useOperaLogin } from './useOperaLogin';
 export { useWebWalletLogin } from './useWebWalletLogin';
 export { useLedgerLogin } from './useLedgerLogin';
 export { useWalletConnectLogin } from './useWalletConnectLogin';